Fix other tests
